Add chat to config GUI - Still in beta version; bugs still exist

+++ core/src/main/java/com/mndk/bteterrarenderer/core/gui/mcfx/button/McFXButton.java
@@ -36,7 +36,7 @@ public class McFXButton extends McFXElement {
     }
 
     @Override
-    public boolean mouseHovered(double mouseX, double mouseY, float partialTicks, boolean mouseHidden) {
+    public boolean mouseHovered(int mouseX, int mouseY, float partialTicks, boolean mouseHidden) {
         return this.button.mouseHovered(mouseX, mouseY, partialTicks, mouseHidden);
     }
